feat(daemon): update the model configs
