+++ crates/daemon/src/config/mod.rs
@@ -10,21 +10,20 @@ pub use default::{
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use std::path::PathBuf;
-pub use {agent::AgentConfig, loader::load_agents_dir};
 pub use {channel::ChannelConfig, mcp::McpServerConfig};
+pub use {loader::load_agents_dir, model::ModelConfig};
 
-mod agent;
 mod default;
 mod loader;
 mod mcp;
+mod model;
 
 /// Top-level daemon configuration.
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Default)]
 pub struct DaemonConfig {
-    /// LLM provider configurations keyed by a user-defined name.
-    /// If a provider's `model` field is empty, it is filled from the key.
+    /// Model configurations.
     #[serde(default)]
-    pub models: BTreeMap<CompactString, ProviderConfig>,
+    pub model: ModelConfig,
     /// Channel configurations keyed by a user-defined name.
     /// If a channel's `platform` field is empty, it is filled from the key.
     #[serde(default)]
@@ -32,9 +31,6 @@ pub struct DaemonConfig {
     /// MCP server configurations.
     #[serde(default)]
     pub mcp_servers: BTreeMap<CompactString, mcp::McpServerConfig>,
-    /// Agent configurations.
-    #[serde(default)]
-    pub agents: AgentConfig,
     /// Optional symlink path for the workspace sandbox root (`~/.walrus/work/`).
     ///
     /// When set, a symlink is created at this path pointing to `~/.walrus/work/`.
@@ -46,11 +42,15 @@ impl DaemonConfig {
     /// Parse a TOML string into a `DaemonConfig`.
     pub fn from_toml(toml_str: &str) -> Result<Self> {
         let mut config: Self = toml::from_str(toml_str)?;
-        config.models.iter_mut().for_each(|(key, provider)| {
-            if provider.model.is_empty() {
-                provider.model = key.clone();
-            }
-        });
+        config
+            .model
+            .providers
+            .iter_mut()
+            .for_each(|(key, provider)| {
+                if provider.model.is_empty() {
+                    provider.model = key.clone();
+                }
+            });
         config.channels.iter_mut().for_each(|(key, channel)| {
             if channel.platform.is_empty() {
                 channel.platform = key.clone();
@@ -68,53 +68,5 @@ impl DaemonConfig {
     pub fn load(path: &std::path::Path) -> Result<Self> {
         let content = std::fs::read_to_string(path)?;
         Self::from_toml(&content)
-    }
-}
-
-#[cfg(not(feature = "local"))]
-impl Default for DaemonConfig {
-    fn default() -> Self {
-        Self {
-            models: [(
-                "deepseek-chat".into(),
-                ProviderConfig {
-                    model: "deepseek-chat".into(),
-                    api_key: None,
-                    base_url: None,
-                    loader: None,
-                    quantization: None,
-                    chat_template: None,
-                },
-            )]
-            .into(),
-            channels: Default::default(),
-            mcp_servers: Default::default(),
-            agents: AgentConfig::default(),
-            work_dir: None,
-        }
-    }
-}
-
-#[cfg(feature = "local")]
-impl Default for DaemonConfig {
-    fn default() -> Self {
-        Self {
-            models: [(
-                "local".into(),
-                ProviderConfig {
-                    model: "Qwen/Qwen3-4B".into(),
-                    api_key: None,
-                    base_url: None,
-                    loader: Some(model::Loader::Text),
-                    quantization: None,
-                    chat_template: None,
-                },
-            )]
-            .into(),
-            channels: Default::default(),
-            mcp_servers: Default::default(),
-            agents: AgentConfig::default(),
-            work_dir: None,
-        }
     }
 }
